Move createquestions command tests to question app
