+++ backend/experiment/management/tests.py
@@ -26,12 +26,5 @@ class CompilePlaylistTest(TestCase):
         finally:        
             remove(filename)  # Make sure csv file is deleted even if tests fail
 
-    def test_createquestions(self):
-        from question.models import Question, QuestionGroup
-        call_command('createquestions')
-        self.assertEqual(len(Question.objects.all()), 161) # Only built-in questions in test database
-        self.assertEqual(len(QuestionGroup.objects.all()), 18) # Only built-in question groups in test database
-        self.assertEqual(len(Question.objects.filter(key='dgf_country_of_origin')), 1)
-        self.assertEqual(len(QuestionGroup.objects.filter(key='DEMOGRAPHICS')), 1)
